Stabilized the ray-based surface tool to work with high-res data.
The critical part of the problem is that the globalElevationRange that
the intersect code uses are the averaged min/maxes. That is to say if
you have high-res of Port-au-Prince the global ranges will be 0,0
because PaP gets averaged out. The real fix is to update construo to
keep track of the min/maxes and store those in the quadtree file headers
for the tiles.