Visibility achieved as on/off switch for pixels.
Caveats: 1. homography fails for segemnts that intersect the nodes but
whos control points are far outside the node's frustum. 2. GL_LINE for
the coverage with just line-width has limitations, need to fiddle with
the linewidth wrt to the view. Might need to also generate these on the
terrain to tweak wrt twist.