Precision issues on the coverage map cornered.
The transform on the GPU is in floating point, thus it would require relative coordinates. For now, I'm just transforming on the CPU. Also need to make the line width relative to the size of the line vs the size of the cells.

Moved the symbol width and length to the settings.