Rolled-back time-stamping of coverage.
the stamping scheme didn't make sense, as it lead to all nodes being updated if one leaf changed (since that propagates a new stamp all the way to the root). Instead, the system counts on the proper update of the ancestry currently cached and simply flags the leaf nodes to propagate the changes down (as was done before).